requirements.txt
langgraph>=0.2
langchain>=0.2
langchain-core>=0.2
langchain-groq>=0.2
langchain-community>=0.2
pydantic>=2.0
pydantic-settings>=2.0
tavily-python>=0.3
streamlit>=1.35
python-slugify>=8.0
python-dotenv>=1.0
tenacity>=8.2
requests>=2.31
pytest>=8.0
duckduckgo-search>=6.0
ddgs>=9.13.1
